Moved RandomColor into the model directory
